src/core/automation.py
import asyncio
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.action_chains import ActionChains
from web3 import Web3
from .fingerprint import generate_fingerprint, apply_fingerprint
from .proxy_manager import ProxyManager
from .sybil_defender import SybilDefender
from .utils import random_delay, async_random_delay
from .tools import log, handle_error

async def automate_airdrop(user_id: str, proxy_manager: ProxyManager, sybil_defender: SybilDefender):
    """异步执行空投任务，优化交互自然性"""
    try:
        await asyncio.to_thread(sybil_defender.check_sybil)
        proxy = proxy_manager.proxies[random.randint(0, len(proxy_manager.proxies) - 1)]
        fingerprint = generate_fingerprint()

        chrome_options = webdriver.ChromeOptions()
        chrome_options.add_argument(f"--proxy-server={proxy}")
        driver = webdriver.Chrome(options=chrome_options)
        apply_fingerprint(driver, fingerprint)

        # 访问空投网站
        driver.get("https://airdrop.example.com")
        await async_random_delay(1, 3)

        # 填写表单（多样化任务，防止 Sybil 检测）
        email = f"user{random.randint(1, 1000)}@example.com"
        driver.find_element(By.ID, "email").send_keys(email)
        
        # 模拟 DeFi 交互（参考 CryptoMelon365 的建议）
        defi_tasks = [
            {"contract": "STG", "action": "deposit", "url": "https://stg.example.com/deposit"},
            {"contract": "Zerolend", "action": "swap", "url": "https://zerolend.example.com/swap"},
            {"contract": "odos", "action": "lend", "url": "https://odos.example.com/lend"}
        ]
        for task in random.sample(defi_tasks, k=random.randint(1, len(defi_tasks))):
            driver.get(task["url"])
            try:
                driver.find_element(By.ID, f"{task['action']}_btn").click()
            except Exception as e:
                log(f"Failed to perform {task['action']} on {task['contract']}: {str(e)}")
            await async_random_delay(2, 5)

        # 提交表单
        driver.find_element(By.ID, "submit").click()

        # 连接钱包并签名（使用硬件钱包或安全存储）
        w3 = Web3(Web3.HTTPProvider("https://rpc.example.com"))
        account = w3.eth.account.from_key("your_private_key")  # 替换为安全存储
        tx = {
            "to": "airdrop_contract_address",
            "value": 0,
            "gas": 21000,
            "nonce": w3.eth.get_transaction_count(account.address),
            "chainId": 1  # 根据网络调整
        }
        signed_tx = w3.eth.account.sign_transaction(tx, account.privateKey)
        tx_hash = w3.eth.send_raw_transaction(signed_tx.rawTransaction)

        # 模拟更自然的交互
        actions = ActionChains(driver)
        for _ in range(random.randint(2, 5)):
            actions.move_by_offset(random.randint(10, 100), random.randint(10, 100)).perform()
            await async_random_delay(0.5, 2)

        log(f"Airdrop task completed with TX: {tx_hash.hex()}")
    except Exception as e:
        handle_error(e, "Airdrop task")
    finally:
        driver.quit()
        sybil_defender.instance_count -= 1

async def run_multiple_airdrops(user_id: str, proxy_manager: ProxyManager, sybil_defender: SybilDefender, num_instances: int = 5):
    """异步运行多个空投任务，限制并发以防 Sybil 检测"""
    semaphore = asyncio.Semaphore(sybil_defender.max_instances)  # 限制并发实例数
    async def limited_airdrop():
        async with semaphore:
            await automate_airdrop(user_id, proxy_manager, sybil_defender)

    tasks = [limited_airdrop() for _ in range(num_instances)]
    try:
        await asyncio.gather(*tasks)
        log(f"Completed {num_instances} airdrop tasks")
    except Exception as e:
        handle_error(e, "Running multiple airdrops")

if __name__ == "__main__":
    proxy_manager = ProxyManager()
    asyncio.run(proxy_manager.batch_import_proxies("proxies.csv"))
    sybil_defender = SybilDefender("user@example.com")
    asyncio.run(run_multiple_airdrops("user@example.com", proxy_manager, sybil_defender, 5))